pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ateam-cli"
version = "0.1.0"
description = "Production-grade, security-hardened multi-agent AI orchestration tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "A-Team Contributors", email = "support@ateam.dev" }
]
keywords = [
    "ai",
    "multi-agent",
    "orchestration",
    "cli",
    "gemini",
    "claude",
    "gpt",
    "ollama",
    "mcp"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # CLI Framework
    "typer>=0.12.0",
    "rich>=13.7.0",
    
    # Data Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    
    # Security
    "keyring>=25.0.0",
    
    # HTTP Client
    "httpx>=0.27.0",
    
    # Configuration
    "pyyaml>=6.0.1",
    
    # AI Provider SDKs
    "google-generativeai>=0.4.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    
    # Code Quality
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    
    # Security Auditing
    "bandit>=1.7.7",
    "pip-audit>=2.7.0",
    "safety>=3.0.0",
    
    # Type Stubs
    "types-pyyaml>=6.0.12",
]

[project.scripts]
ateam = "ateam.cli.main:app"

[project.urls]
Homepage = "https://github.com/ghassan-gaidi/A-Team-CLI"
Documentation = "https://docs.ateam.dev"
Repository = "https://github.com/ghassan-gaidi/A-Team-CLI"
Issues = "https://github.com/ghassan-gaidi/A-Team-CLI/issues"

# ============================================================================
# TOOL CONFIGURATION
# ============================================================================

[tool.hatch.build.targets.wheel]
packages = ["ateam"]

# Ruff - Fast Python linter and formatter
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # use of assert (OK in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]  # Allow assert and unused args in tests

# MyPy - Static type checker
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.generativeai.*",
    "anthropic.*",
    "keyring.*",
]
ignore_missing_imports = true

# Pytest - Testing framework
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=ateam",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "security: Security tests",
    "slow: Slow tests",
]
asyncio_mode = "auto"

# Coverage - Code coverage reporting
[tool.coverage.run]
source = ["ateam"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit - Security linter
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]  # Allow assert in tests
